requirements.txt
absl-py==0.9.0
argon2-cffi==20.1.0
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
blis==0.4.1
bokeh==2.3.0
boto==2.49.0
boto3==1.14.48
botocore==1.17.48
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
hyperlink==19.0.0
idna==2.10
imageio==2.9.0
incremental==17.5.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itemadapter==0.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
Keras==2.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.2
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
networkx==2.4
#-e git+https://github.com/huggingface/neuralcoref.git@18c0f4c5c4d50aa6235470f6b2b29feae72bee9b#egg=neuralcoref
nltk==3.5
notebook==6.1.4
numpy==1.18.5
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.1
#-e git+https://github.com/pdfminer/pdfminer.six.git@0b44f7771462363528c109f263276eb254c4fcd0#egg=pdfminer.six
pdftotree==0.4.0
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.12.4
ptyprocess==0.6.0
py==1.9.0
pyahocorasick==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pytest==6.1.2
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
queuelib==1.5.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
Scrapy==2.2.1
seaborn==0.11.1
selectivesearch==0.4
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.91
service-identity==18.1.0
six==1.15.0
sklearn==0.0
smart-open==2.1.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
tabula-py==2.1.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2020.7.24
tokenizers==0.8.1rc2
toml==0.10.1
torch==1.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==3.1.0
Twisted==20.3.0
typing-extensions==3.7.4.3
urllib3==1.25.10
w3lib==1.22.0
Wand==0.6.2
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zope.interface==5.1.0
